+++ codex/urls/opds/authentication.py
@@ -3,9 +3,8 @@
 from django.urls import path
 from django.views.decorators.cache import cache_page
 
+from codex.urls.const import COMMON_TIMEOUT
 from codex.views.opds.authentication_v1 import OPDSAuthentication1View
-
-TIMEOUT = 60 * 60
 
 app_name = "authentication"
 
@@ -13,7 +12,7 @@ app_name = "authentication"
 urlpatterns = [
     path(
         "v1",
-        cache_page(TIMEOUT)(OPDSAuthentication1View.as_view()),
+        cache_page(COMMON_TIMEOUT)(OPDSAuthentication1View.as_view()),
         name="v1",
     ),
 ]