+++ codex/urls/root.py
@@ -9,8 +9,6 @@ from django.views.generic.base import RedirectView
 
 from codex.urls.converters import GroupConverter
 
-TIMEOUT = 60 * 60
-
 register_converter(GroupConverter, "group")
 
 urlpatterns = [