+++ frontend/src/components/reader/drawer/download-panel.vue
@@ -13,19 +13,13 @@
         >Download Book
       </v-list-item-title>
     </v-list-item>
-    <v-list-item v-if="isPDF" @click="openBookInBrowser">
-      <v-list-item-title>
-        <v-icon>{{ mdiOpenInNew }}</v-icon
-        >Read PDF with Browser
-      </v-list-item-title>
-    </v-list-item>
   </div>
 </template>
+
 <script>
-import { mdiDownload, mdiFileImage, mdiOpenInNew } from "@mdi/js";
+import { mdiDownload, mdiFileImage } from "@mdi/js";
 import { mapActions, mapGetters, mapState } from "pinia";
 
-import { getBookInBrowserURL } from "@/api/v3/common";
 import { getDownloadPageURL, getDownloadURL } from "@/api/v3/reader";
 import { useCommonStore } from "@/stores/common";
 import { useReaderStore } from "@/stores/reader";
@@ -36,26 +30,23 @@ export default {
     return {
       mdiDownload,
       mdiFileImage,
-      mdiOpenInNew,
     };
   },
   computed: {
     ...mapGetters(useReaderStore, ["activeTitle", "routeParams"]),
     ...mapState(useReaderStore, {
+      currentBook: (state) => state.books?.current,
       storePage: (state) => state.page,
-      filename: (state) => state.books?.current?.filename,
-      downloadURL(state) {
-        return getDownloadURL(state.books?.current?.pk);
-      },
-      bookInBrowserURL(state) {
-        return getBookInBrowserURL(state.books?.current?.pk);
-      },
-      isPDF(state) {
-        return state.books?.current?.fileType == "PDF";
-      },
     }),
+    downloadURL() {
+      return getDownloadURL(this.currentBook);
+    },
+    filename() {
+      return this.currentBook?.filename;
+    },
     pageSrc() {
-      return getDownloadPageURL(this.routeParams);
+      const { pk, mtime } = this.currentBook;
+      return getDownloadPageURL({ pk, page: this.storePage, mtime });
     },
     pageName() {
       return `${this.activeTitle} - page ${this.storePage}.jpg`;
@@ -69,12 +60,10 @@ export default {
     downloadBook() {
       this.downloadIOSPWAFix(this.downloadURL, this.filename);
     },
-    openBookInBrowser() {
-      window.open(this.bookInBrowserURL, "_blank");
-    },
   },
 };
 </script>
+
 <style scoped lang="scss">
 #downloadPanel {
   background-color: rgb(var(--v-theme-background));